Extract "match" assertion to own file
