+++ lib/referee.js
@@ -8,7 +8,6 @@ var captureException = require("./capture-exception");
 var interpolatePosArg = require("./interpolate-pos-arg");
 var interpolateProperties = require("./interpolate-properties");
 var expect = require("./expect");
-var samsam = require("samsam");
 var bane = require("bane");
 var includes = require("lodash.includes");
 var isArguments = require("lodash.isarguments");
@@ -195,51 +194,7 @@ require("./assertions/greater")(referee);
 require("./assertions/less")(referee);
 require("./assertions/defined")(referee);
 require("./assertions/is-null")(referee);
-
-referee.match = function (actual, matcher) {
-    try {
-        return samsam.match(actual, matcher);
-    } catch (e) {
-        throw new Error("Matcher (" + referee.format(matcher) +
-                        ") was not a string, a number, a function, " +
-                        "a boolean or an object");
-    }
-};
-
-referee.add("match", {
-    assert: function (actual, matcher) {
-        var passed;
-
-        try {
-            passed = referee.match(actual, matcher);
-        } catch (e) {
-            this.exceptionMessage = e.message;
-            return this.fail("exceptionMessage");
-        }
-
-        return passed;
-    },
-
-    refute: function (actual, matcher) {
-        var passed;
-
-        try {
-            passed = referee.match(actual, matcher);
-        } catch (e) {
-            this.exceptionMessage = e.message;
-            return this.fail("exceptionMessage");
-        }
-
-        return !passed;
-    },
-
-    assertMessage: "${customMessage}${actual} expected to match ${expected}",
-    refuteMessage: "${customMessage}${actual} expected not to match ${expected}",
-    expectation: "toMatch",
-    values: actualAndExpectedMessageValues
-});
-
-referee.assert.match.exceptionMessage = referee.refute.match.exceptionMessage = "${customMessage}${exceptionMessage}";
+require("./assertions/match")(referee);
 
 referee.add("isObject", {
     assert: function (actual) {
